+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @qualcomm-ui/esbuild Changelog
 
+## 1.1.0
+
+Jun 3rd, 2026
+
+### Features
+
+- [plugin]: redesign library-entries-plugin to support built-in entry discovery ([f600cc7](https://github.com/qualcomm/qualcomm-ui-utils/commit/f600cc7))
+
+### Performance Improvements
+
+- [plugin]: only re-emit changed files in watch mode ([51484d4](https://github.com/qualcomm/qualcomm-ui-utils/commit/51484d4))
+
 ## 1.0.0 (2026/06/03)
 
 - add `libraryEntriesPlugin` and related utilities for building and watching libraries from secondary entry points